Adding Fletching Factory and Spectral Arrows. Also, issues
Fletching factory added. To get it, you upgrade from a laboratory with 512 arrows, 256 eggs, 128 hay bales, 64 nether warts, 32 oak saplings, 32 spruce saplings, and 4 water buckets (which turn into buckets after running the factory). 

The reasoning behind this recipe is that these things all have sticky combinations, and nether wart is useful for potions anyway. Also, the materials come from different shards.

The Spectral Arrows come from inputting 16 arrows and 16 glowstone dust to create 64 spectral arrows. In vanilla minecraft, you can create 32 spectral arrows using just 16 arrows and a 64 glowstone dust. The factory recipe is more efficient by doubling output and cutting glowstone to a quarter of the cost. 

Now for the issues. There does not appear to be a way to create potion tipped arrows with factory mod. In fact, Factory Mod does not seem to recognize potions either. That means that inputs and outputs seem to not have a way to understand the input and the output. If durability can help the factory understand potions and potion tipped arrows (like it does with dyes and saplings) in the future (if not already done so), then we can make some work on the potion arrows.

(For later recipe reference, one lingering potion can create 8 potion arrows in vanilla Minecraft. I recommend that the factory produce more arrows than that. Like 16.)

Anyhow, this should be a good start for creating special arrows. Let us know how we can make the factories understand potion tipped arrows and potions and things can move forward.