Ruby Enrichment balances
Our town runs this recipe and we have a lot of data. The fish
requirement is out of whack with the pufferfish in the Topaz recipe.
Pufferfish are about half the chance of being caught of salmon. So it
seems fair that salmon amounts are doubled, not quadroupled.

Sapplings require really huge farms to get large amounts. It seems fair
to halve sappling amounts on all recipes.

For the rest this recipe seems fine. Vines cost iron to farm but this
can be mitigated by unbreaking shears.