Amethys Balances
This recipe is worthless IMO. No one is going to take this path. It
requires 3 types of tree farms acacia, dark oak/oak for apples and
jungle. On top of that it takes ink.

I've reduced the amounts of hard to get items like apples and ink. But I
don't think it's enough to make this viable. Either oak or ink needs to
be changed to an easier to farm ingredient. Ink is already needed for
bastions as well, so I suggest that part be changed.