Introducing the slow stone smelter
The slow stone smelter is a factory that converts 3/4 of a double chest of cobblestone into a double chest of stone over the course of 24 hours. It makes 108 cobblestone into 144 stone each hour.

The set-up cost is 2.5 stacks of cobblestone and 2.5 stacks of stone, on top of the basic contraption stuff. I would prefer to skip the basic contraption and possibly increase the stone and cobblestone cost, but I am guessing the *other developers* feel strongly about using the basic contraption as a starting point.

The repair cost is 2 essence which should allow one person to run around 35 of them if he really wanted to.

The charcoal cost should be around 64 per day but honestly I am not sure. I only put it that low to allow it to be left unattended for 24 hours. It should be raised if another viable way to achieve that is added, like hoppers or more compact fuel.

Note that I am proposing this in anticipation of updates in the near future that will make it undesirable to rebuild factories to sidestep the essence upkeep cost.